Static cast T in double before calling sqrt in Statistics<T>.
Otherwise vcxx complains about an ambiguity when T is neither double nor
float.

	* include/libport/statistics.hxx (standard_deviation): There.

+++ include/libport/statistics.hxx
@@ -56,7 +56,7 @@ namespace libport
   inline T
   Statistics<T>::standard_deviation() const
   {
-    return static_cast<T>(sqrt(variance()));
+    return static_cast<T>(sqrt(static_cast<double>(variance())));
   }
 
   template<typename T>
